# packages/stock-data/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GlobalStockDataService",
    "description": "面向股票分析、量化交易与策略回测的全球股票数据服务 (50GB空间极致优化架构)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/market/margin": {
      "get": {
        "tags": [
          "Margin Trading (融资融券)"
        ],
        "summary": "Get Market Margin",
        "description": "获取全市场融资融券每日走势 (做多/做空压力标准指标):\n融资余额、融资买入额、融券余量/余额、融资融券余额合计。",
        "operationId": "get_market_margin_api_v1_market_margin_get",
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(SH|SZ)$",
              "description": "市场: SH (上交所) 或 SZ (深交所)。北交所暂无两融业务，不支持",
              "title": "Market"
            },
            "description": "市场: SH (上交所) 或 SZ (深交所)。北交所暂无两融业务，不支持"
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "开始日期 YYYY-MM-DD (默认近30日)",
              "title": "Start"
            },
            "description": "开始日期 YYYY-MM-DD (默认近30日)"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "结束日期 YYYY-MM-DD (默认当天)",
              "title": "End"
            },
            "description": "结束日期 YYYY-MM-DD (默认当天)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarginMarketResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stock/margin": {
      "get": {
        "tags": [
          "Margin Trading (融资融券)"
        ],
        "summary": "Get Stock Margin",
        "description": "获取个股融资融券明细 (仅两融标的池内股票支持):\n融资买入额/余额、融券卖出量/余量。北交所无两融业务，一律返回未找到。",
        "operationId": "get_stock_margin_api_v1_stock_margin_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "股票代码，如 600519 (茅台，需为两融标的), 000001 (平安银行)",
              "title": "Symbol"
            },
            "description": "股票代码，如 600519 (茅台，需为两融标的), 000001 (平安银行)"
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "查询日期 YYYY-MM-DD (默认最近一个交易日，交易所两融明细通常 T+1 披露)",
              "title": "Date"
            },
            "description": "查询日期 YYYY-MM-DD (默认最近一个交易日，交易所两融明细通常 T+1 披露)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarginStockResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarginMarketItem": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date",
            "description": "信用交易日期: YYYY-MM-DD"
          },
          "financing_buy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Financing Buy",
            "description": "融资买入额 (元)"
          },
          "financing_balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Financing Balance",
            "description": "融资余额 (元)"
          },
          "securities_lending_volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Securities Lending Volume",
            "description": "融券卖出量/余量 (股)"
          },
          "securities_lending_balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Securities Lending Balance",
            "description": "融券余额 (元)"
          },
          "total_balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Balance",
            "description": "融资融券余额合计 (元)"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "MarginMarketItem"
      },
      "MarginMarketResponse": {
        "properties": {
          "market": {
            "type": "string",
            "title": "Market"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/MarginMarketItem"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "market",
          "count",
          "data"
        ],
        "title": "MarginMarketResponse"
      },
      "MarginStockItem": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date",
            "description": "信用交易日期: YYYY-MM-DD"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "financing_buy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Financing Buy",
            "description": "融资买入额 (元)"
          },
          "financing_balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Financing Balance",
            "description": "融资余额 (元)"
          },
          "lending_sell_volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lending Sell Volume",
            "description": "融券卖出量 (股)"
          },
          "lending_balance_volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lending Balance Volume",
            "description": "融券余量 (股)"
          },
          "total_balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Balance",
            "description": "融资融券余额合计 (元)"
          }
        },
        "type": "object",
        "required": [
          "date",
          "symbol"
        ],
        "title": "MarginStockItem"
      },
      "MarginStockResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "found": {
            "type": "boolean",
            "title": "Found"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MarginStockItem"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "date",
          "found"
        ],
        "title": "MarginStockResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}